修改模板更新，删除接口，对于有任务使用的模板，不能删除，不能修改

+++ blade-service/mk-composition-config/src/main/java/org/springblade/composition/controller/TemplateController.java
@@ -41,6 +41,8 @@ import org.springblade.core.tool.api.R;
 import org.springblade.core.tool.api.ResultCode;
 import org.springblade.core.tool.utils.BeanUtil;
 import org.springblade.flow.core.feign.IFlowEngineClient;
+import org.springblade.task.entity.Task;
+import org.springblade.task.feign.ITaskClient;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
@@ -70,6 +72,7 @@ public class TemplateController extends BladeController {
 	private final ITemplateCompositionService templateCompositionService;
 	private final ICompositionService compositionService;
 	private final IFlowEngineClient iFlowEngineClient;
+	private final ITaskClient taskClient;
 
 	@GetMapping("/role-field")
 	@ApiOperationSupport(order = 8)
@@ -170,9 +173,15 @@ public class TemplateController extends BladeController {
 	@Transactional(rollbackFor = Exception.class)
 	@ApiOperation(value = "新增或修改模板，同时构建组合", notes = "传入templateDTO")
 	public R submit(@Valid @RequestBody TemplateDTO templateDTO) {
-//		List<TemplateComposition> templateCompositions = templateDTO.getTemplateCompositions();
-//		templateCompositions.forEach(templateComposition -> templateComposition.setTemplateId(template.getId()));
-//		return R.status(templateService.compose(templateCompositions));
+		Template template = Objects.requireNonNull(BeanUtil.copy(templateDTO, Template.class));
+		// 不等于null，就是更新操作。需要首先判断是否有任务已经使用此模板
+		if (template.getId() != null){
+			Task task = taskClient.getByTemplate(template.getId()).getData();
+			if (task.getId() != null){
+				return R.fail("有任务已经使用此模板，不能修改，可以停用");
+			}
+		}
+
 		List<TemplateCompositionDTO> templateCompositions = templateDTO.getTemplateCompositions();
 		templateCompositions.forEach(templateCompositionDTO -> {
 			Composition composition = compositionService.getById(templateCompositionDTO.getCompositionId());
@@ -185,7 +194,6 @@ public class TemplateController extends BladeController {
 			return R.fail("部署模版流程失败");
 		}
 		templateDTO.setProcessDefinitionId((String)result.getData());
-		Template template = Objects.requireNonNull(BeanUtil.copy(templateDTO, Template.class));
 		boolean tmp = templateService.saveOrUpdate(template);
 		if(!tmp) {
 			return R.fail("新建模板失败");
@@ -196,6 +204,7 @@ public class TemplateController extends BladeController {
 			return templateComposition;
 		}).collect(Collectors.toList());
 		return R.status(templateService.compose(templateCompositionList));
+
 	}
 
 
@@ -207,6 +216,10 @@ public class TemplateController extends BladeController {
 	@ApiOperationSupport(order = 4)
 	@ApiOperation(value = "删除", notes = "传入ids")
 	public R remove(@ApiParam(value = "主键集合", required = true) @RequestParam String ids) {
+		Task task = taskClient.getByTemplate(Long.valueOf(ids)).getData();
+		if (task.getId() != null){
+			return R.fail("有任务已经使用此模板，只能停用，不能删除");
+		}
 		return R.status(templateService.remove(ids));
 	}
